+++ src/RhinoInside.Revit.GH/Types/ObjectStyles/FillPatternElement.cs
@@ -1,12 +1,20 @@
 using System;
+using System.Linq;
 using System.Collections.Generic;
 using Grasshopper.Kernel;
 using Rhino;
+using Rhino.Geometry;
 using Rhino.DocObjects;
 using ARDB = Autodesk.Revit.DB;
+#if RHINO_8
+using Grasshopper.Rhinoceros.Drafting;
+using Grasshopper.Rhinoceros;
+#endif
 
 namespace RhinoInside.Revit.GH.Types
 {
+  using Convert.Units;
+
   [Kernel.Attributes.Name("Fill Pattern")]
   public class FillPatternElement : Element, Bake.IGH_BakeAwareElement
   {
@@ -21,6 +29,21 @@ namespace RhinoInside.Revit.GH.Types
     {
       return Element.FromElementId(doc, id) as FillPatternElement;
     }
+
+    #region IGH_Goo
+    public override bool CastTo<Q>(out Q target)
+    {
+#if RHINO_8
+      if (typeof(Q).IsAssignableFrom(typeof(ModelHatchPattern)))
+      {
+        target = (Q) (object) ToModelContent(new Dictionary<ARDB.ElementId, ModelContent>());
+        return true;
+      }
+#endif
+
+      return base.CastTo(out target);
+    }
+    #endregion
 
     #region IGH_BakeAwareElement
     bool IGH_BakeAwareData.BakeGeometry(RhinoDoc doc, ObjectAttributes att, out Guid guid) =>
@@ -43,12 +66,6 @@ namespace RhinoInside.Revit.GH.Types
       {
         using (var pattern = fillPattern.GetFillPattern())
         {
-          if (pattern.IsSolidFill)
-          {
-            idMap.Add(Id, guid = HatchPattern.Defaults.Solid.Id);
-            return true;
-          }
-
           // 2. Check if already exist
           var hatchPattern = doc.HatchPatterns.FindName(fillPattern.Name) ??
                              new HatchPattern() { Name = fillPattern.Name };
@@ -58,17 +75,49 @@ namespace RhinoInside.Revit.GH.Types
           // 3. Update if necessary
           if (index < 0 || overwrite)
           {
-            // TODO: No API
-            //hatchPattern.FillType = HatchPatternFillType.Lines;
+            if (pattern.IsSolidFill)
+            {
+              hatchPattern.FillType = HatchPatternFillType.Solid;
+              hatchPattern.Name = "<Solid fill>";
+            }
+            else
+            {
+              hatchPattern.FillType = HatchPatternFillType.Lines;
 
-            hatchPattern = pattern.GridCount == 2 ?
-              HatchPattern.Defaults.Grid:
-              HatchPattern.Defaults.Hatch1;
-            hatchPattern.Name = fillPattern.Name;
+#if RHINO_8
+              var grids = pattern.GetFillGrids();
+              for (int g = 0; g < grids.Count; ++g)
+              {
+                var grid = grids[g];
+                var line = new HatchLine()
+                {
+                  BasePoint = new Point2d(UnitScale.Convert(grid.Origin.U, UnitScale.Internal, UnitScale.Millimeters), UnitScale.Convert(grid.Origin.V, UnitScale.Internal, UnitScale.Millimeters)),
+                  Offset = new Vector2d(UnitScale.Convert(grid.Shift, UnitScale.Internal, UnitScale.Millimeters), UnitScale.Convert(grid.Offset, UnitScale.Internal, UnitScale.Millimeters)),
+                  Angle = grid.Angle,
+                };
+                line.SetDashes(grid.GetSegments().Select((x, i) => UnitScale.Convert(i % 2 == 0 ? +x : -x, UnitScale.Internal, UnitScale.Millimeters)));
+
+                hatchPattern.AddHatchLine(line);
+              }
+#else
+              hatchPattern = pattern.GridCount == 2 ?
+                HatchPattern.Defaults.Grid :
+                HatchPattern.Defaults.Hatch1;
+
+              hatchPattern.Index = -1;
+              hatchPattern.Id = Guid.Empty;
+              hatchPattern.Name = fillPattern.Name;
+#endif
+            }
           }
 
+#if RHINO_8
+          if (index < 0) { index = doc.HatchPatterns.Add(hatchPattern); hatchPattern = doc.HatchPatterns[index]; }
+          else if (overwrite) { doc.HatchPatterns.Modify(hatchPattern, index, true); }
+#else
           if (index < 0) { index = doc.HatchPatterns.Add(hatchPattern); hatchPattern = doc.HatchPatterns[index]; }
           else if (overwrite) { /*doc.HatchPatterns.Modify(hatchPattern, index, true);*/ }
+#endif
 
           idMap.Add(Id, guid = hatchPattern.Id);
           return true;
@@ -77,6 +126,50 @@ namespace RhinoInside.Revit.GH.Types
 
       return false;
     }
+    #endregion
+
+    #region ModelContent
+#if RHINO_8
+    internal ModelContent ToModelContent(IDictionary<ARDB.ElementId, ModelContent> idMap)
+    {
+      if (idMap.TryGetValue(Id, out var modelContent))
+        return modelContent;
+
+      if (Value is ARDB.FillPatternElement fillPatter)
+      {
+        var attributes = new ModelHatchPattern.Attributes() { Path = fillPatter.Name };
+        using (var pattern = fillPatter.GetFillPattern())
+        {
+          if (pattern.IsSolidFill)
+          {
+            attributes.Name = "<Solid fill>";
+          }
+          else
+          {
+            var grids = pattern.GetFillGrids();
+            var lines = new ModelHatchLine[grids.Count];
+            for(int g = 0; g < grids.Count; ++g)
+            {
+              var grid = grids[g];
+              lines[g] = new ModelHatchLine.Attributes
+              {
+                Base = new Point2d(UnitScale.Convert(grid.Origin.U, UnitScale.Internal, UnitScale.Millimeters), UnitScale.Convert(grid.Origin.V, UnitScale.Internal, UnitScale.Millimeters)),
+                Offset = new Vector2d(UnitScale.Convert(grid.Shift, UnitScale.Internal, UnitScale.Millimeters), UnitScale.Convert(grid.Offset, UnitScale.Internal, UnitScale.Millimeters)),
+                Angle = grid.Angle,
+                Segments = grid.GetSegments().Select((x, i) => UnitScale.Convert(i % 2 == 0 ? +x : -x, UnitScale.Internal, UnitScale.Millimeters)).ToArray()
+              };
+            }
+            attributes.Lines = lines;
+          }
+        }
+
+        idMap.Add(Id, modelContent = attributes.ToModelData() as ModelContent);
+        return modelContent;
+      }
+
+      return null;
+    }
+#endif
     #endregion
   }
 }